Creates Admin Context to keep middleware more organized
