+++ src/context/AuthenticationContext.js
@@ -17,86 +17,6 @@ const reducer = (state, action) => {
   }
 };
 
-
-//======================= USERS =======================//
-// Register
-const registerUser = dispatch => async (data) => {
-  const REGISTER_ENDPOINT = 'http://localhost:8080/con/api/users/register.php';
-  dispatch({ type: 'reset_error' });
-  dispatch({ type: 'reset_success' });
-  dispatch({ type: 'start_loading' });
-
-  try {
-    const response = await axios({
-      method: 'post',
-      url: REGISTER_ENDPOINT,
-      headers: {
-        'content-type': 'application/json'
-      },
-      data: data
-    });
-    dispatch({ type: 'stop_loading' });
-    dispatch({ type: 'set_success', payload: 'registerUser' });
-    return response;
-  }
-  catch (e) {
-    dispatch({ type: 'stop_loading' });
-    dispatch({ type: 'set_error', payload: 'registerUser' });
-  }
-};
-
-// Promote
-const promote = dispatch => async (data) => {
-  const PROMOTE_ENDPOINT = 'http://localhost:8080/con/api/users/promote.php';
-  dispatch({ type: 'reset_error' });
-  dispatch({ type: 'reset_success' });
-  dispatch({ type: 'start_loading' });
-
-  try {
-    const response = await axios({
-      method: 'put',
-      url: PROMOTE_ENDPOINT,
-      headers: {
-        'content-type': 'application/json'
-      },
-      data: data
-    });
-    dispatch({ type: 'stop_loading' });
-    dispatch({ type: 'set_success', payload: 'promotionUser' });
-    return response;
-  }
-  catch (e) {
-    dispatch({ type: 'stop_loading' });
-    dispatch({ type: 'set_error', payload: 'promotionUser' });
-  }
-};
-
-// Delete
-const deleteUser = dispatch => async (data) => {
-  const DELETE_ENDPOINT = 'http://localhost:8080/con/api/users/delete.php';
-  dispatch({ type: 'reset_error' });
-  dispatch({ type: 'reset_success' });
-  dispatch({ type: 'start_loading' });
-
-  try {
-    const response = await axios({
-      method: 'delete',
-      url: DELETE_ENDPOINT,
-      headers: {
-        'content-type': 'application/json'
-      },
-      data: data
-    });
-    dispatch({ type: 'stop_loading' });
-    dispatch({ type: 'set_success', payload: 'deleteUser' });
-    return response;
-  }
-  catch (e) {
-    dispatch({ type: 'stop_loading' });
-    dispatch({ type: 'set_error', payload: 'deleteUser' });
-  }
-};
-
 // Sign In
 const signin = dispatch => async ({ email, password }) => {
   const LOGIN_ENDPOINT = 'http://localhost:8080/con/api/users/login.php'
@@ -146,118 +66,6 @@ const signout = dispatch => async () => {
   localStorage.removeItem('user');
 };
 
-//======================= GROUPS =======================//
-// Delete Group
-
-const deleteGroup = dispatch => async (data) => {
-  const DELETE_ENDPOINT = 'http://localhost:8080/con/api/groups/delete.php';
-  dispatch({ type: 'reset_error' });
-  dispatch({ type: 'reset_success' });
-  dispatch({ type: 'start_loading' });
-
-  try {
-    const response = await axios({
-      method: 'delete',
-      url: DELETE_ENDPOINT,
-      headers: {
-        'content-type': 'application/json'
-      },
-      data: data
-    });
-    dispatch({ type: 'stop_loading' });
-    dispatch({ type: 'set_success', payload: 'deleteGroup' });
-    return response;
-  }
-  catch (e) {
-    dispatch({ type: 'stop_loading' });
-    dispatch({ type: 'set_error', payload: 'deleteGroup' });
-  }
-};
-
-//======================= CONDO ASSOCIATIONS =======================//
-
-// Register Condo Association
-
-const registerCA = dispatch => async (data) => {
-  const REGISTER_ENDPOINT = 'http://localhost:8080/con/api/associations/register.php';
-  dispatch({ type: 'reset_error' });
-  dispatch({ type: 'reset_success' });
-  dispatch({ type: 'start_loading' });
-
-  try {
-    const response = await axios({
-      method: 'post',
-      url: REGISTER_ENDPOINT,
-      headers: {
-        'content-type': 'application/json'
-      },
-      data: data
-    });
-    dispatch({ type: 'stop_loading' });
-    dispatch({ type: 'set_success', payload: 'registerCA' });
-    return response;
-  }
-  catch (e) {
-    dispatch({ type: 'stop_loading' });
-    dispatch({ type: 'set_error', payload: 'registerCA' });
-  }
-};
-
-// Register User to Condo Association
-
-const assignUser = dispatch => async (data) => {
-  const ASSIGNMENT_ENDPOINT = 'http://localhost:8080/con/api/associations/assign.php';
-  dispatch({ type: 'reset_error' });
-  dispatch({ type: 'reset_success' });
-  dispatch({ type: 'start_loading' });
-
-  try {
-    const response = await axios({
-      method: 'post',
-      url: ASSIGNMENT_ENDPOINT,
-      headers: {
-        'content-type': 'application/json'
-      },
-      data: data
-    });
-    dispatch({ type: 'stop_loading' });
-    dispatch({ type: 'set_success', payload: 'assignCA' });
-    return response;
-  }
-  catch (e) {
-    dispatch({ type: 'stop_loading' });
-    dispatch({ type: 'set_error', payload: 'assignCA' });
-  }
-};
-
-// Delete Condo Association
-
-const deleteCA = dispatch => async (data) => {
-  const DELETE_ENDPOINT = 'http://localhost:8080/con/api/associations/delete.php';
-  dispatch({ type: 'reset_error' });
-  dispatch({ type: 'reset_success' });
-  dispatch({ type: 'start_loading' });
-
-  try {
-    const response = await axios({
-      method: 'delete',
-      url: DELETE_ENDPOINT,
-      headers: {
-        'content-type': 'application/json'
-      },
-      data: data
-    });
-    dispatch({ type: 'stop_loading' });
-    dispatch({ type: 'set_success', payload: 'deleteCA' });
-    return response;
-  }
-  catch (e) {
-    dispatch({ type: 'stop_loading' });
-    dispatch({ type: 'set_error', payload: 'deleteCA' });
-  }
-};
-
-
 export const { Context, Provider } = createDataContext(reducer, {
-  registerUser, promote, deleteUser, signin, editProfile, deleteGroup, registerCA, assignUser, deleteCA, signout
+  signin, editProfile, signout
 }, { isLoading: false, error: '', success: '', user: null });
